chore: Extract reusable device-related code and add small improvements (#6060)
There was quite a bit of code related to devices in guest_vm_runner that
would be useful in other codepaths so this PR extracts the code out to a
separate crate.

I've also found that when we attach a disk image to a loop device, there
is a race condition since finding a free loop device and attaching the
image is not atomic. This PR adds a fix.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>