fix: Log sender reply callback instead of request ID (#6102)
There are two log lines which need to be "linked" via an ID
corresponding to the request.

However one of them logs the callback ID corresponding to the canister
call, while the other logs the ID corresponding to the request from the
management canister. (which is updated
[here](https://github.com/dfinity/ic/blob/e9f4cf612cb649099c2b37f97e761b3362e39cbb/rs/replicated_state/src/metadata_state/subnet_call_context_manager.rs#L244)).

Those IDs are pretty much always corresponding to different requests.