fix(sns): Fix disburse_neuron to prevent overburning fees (#6098)
This PR fixes SNS behavior so that it can't burn fees for proposals that
are still open, in the case when the dissolve delay minimums for
creating proposals don't rule out having open proposals still attached
to the neuron.

Previously, if you opened a proposal, and then dissolved your neuron,
and then disbursed, you would burn the fee associated with the proposal
even if it would have been refunded later. That would result in a
permanent loss of those funds.