feat: Make cost API for threshold keys available on rental subnets (#6125)
The cost API for threshold keys depends on the hosting subnet's size
and, since https://github.com/dfinity/ic/pull/5847, cost schedule.
However, the cost schedule of a different subnet is not known to the
calling subnet. This PR fixes that by adding the cost schedule of each
subnet to the `SubnetTopology`.

A follow-up PR will fix the cost calculation.

---------

Co-authored-by: mraszyk <31483726+mraszyk@users.noreply.github.com>