feat(nns): Created a submit-motion-proposal tool. (#6147)
There are a few preparatory steps that you need to take before running
this tool:

1. Plug in your HSM/Nitro USB device.
2. Unplug your Ubikey (if you have one).
3. Set DFX_HSM_PIN.

This tool has a couple of required flags

1. `--neuron-id`
2. `--proposal-file`

The proposal file has the following format:

```
${HEADER}
--------------------------------------------------------------------------------
${SUMMARY}
```

Notice the line consisting of 80 dash characters separating the two
sections.

`HEADER` in turn is in YAML format, and has two keys:

* `title`
* `url`

As you would expect, upon success, the tools prints out a URL to the now
live motion proposal.