feat: [MR-670] Bump certification version (#6152)
This commit bumps the current certification version to V21. V21
introduces a new format to represent the routing table in the state tree
and was previously implemented in commit 7a809c0.

In addition, this PR also whitelists the new paths for calls to
`api/v2/canister/*/read_state` and `api/v2/subnet/*/read_state`.

This commit is part of the larger canister migration feature and
particularly the effort to expose the routing table in delegations in a
more scalable way.