fix(sns/nns): Fix edge case where simultaneous upgrades can cause deadlock (#6127)
In rare cases, when two upgrade proposals are passed at the same time,
SNS and NNS root can attempt to execute them both at the same time,
causing the canister to become permanently stopped, or to upgrade
unsafely (when it is not fully stopped.

This is fixed by adding a lock to prevent simultaneous upgrade attempts
on the same canister. The second upgrade immediately fails, which can be
recovered by making a subsequent proposal if that is desired.

Queuing and other sequencing mechanisms to try the conflicting upgrade
were considered, but the complexity outweighed any benefit, since this
is a rare edge case.