feat: [MR-665]: Extend canister status endpoint (#6144)
To enable the coming canister migration feature, this PR adds some
additional information to `canister_status`:

- `version` indicates the canister version
- `ready_for_migration` indicates whether the canister is ready for
migration

Instead of adding this bool to the stopped variant, I added it as a
field of `canister_status`.
The former has the advantage that only legal data combinations are
representable with that type. The latter has the advantage that it is a
non-breaking Candid change, which is more important, especially since
there are lots of consumers of this endpoint that we have no control
over.