chore: CON-1549 CON-1550 Report IDKG transcript resolution errors occurring during batch delivery (#6135)
With this PR we increment a critical error counter if the resolution of
an IDKG transcript ref ever fails during batch delivery, which is
unexpected. This requires us to move the batch delivery's `IDkgStats`
into the `idkg` crate.

In order to test the change, we additionally make
`create_available_pre_signature_with_key_transcript` parameterizable by
height. This allows us to test the delivery of a block that contains an
IDKG transcript ref pointing to a block height that doesn't exist. In
that case, the critical error counter should be incremented.

---------

Co-authored-by: Leon Tan <leon.tan@dfinity.org>