chore: CON-1547 Use the full pre-signature to determine "oldest registry version in use" (#6166)
Request contexts are now paired with the full pre-signature (instead of
only the pre-signature ID). Therefore, we can now determine the registry
versions referenced by ongoing signature requests via the transcripts of
the paired pre-signature directly, instead of having to look up the
corresponding full pre-signature in the payload.

The "oldest" such registry version continues to be included in the CUP,
and affects when a node is allowed to leave the subnet. A node may only
leave the subnet if it is no longer part of it according to this
"oldest" registry version. This ensures that all ongoing signature
requests that require the node's contribution may be finished.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>