feat(ICP-Ledger): FI-1552: Implement endpoint to reset legacy approvals (#6121)
The new endpoint allows a principal to remove an approval by specifying
the `from_subaccount` and spender's AccountIdentifier: `remove_approval
: (opt SubAccount, AccountIdentifier) -> (ApproveResult);`. Although the
`from_subaccount` still needs to be specified, the spender's principal
and subaccount are not required. Removing the approval generates a
regular approve block, and a fee is deducted from the caller principal's
`from_subaccount`.

---------

Co-authored-by: Maciej Dfinity <maciej@dfinity.org>