feat: CON-1532 CON-1534 Implement pre-signature stash behind a feature flag (#6179)
This PR adds a new feature flag `store_pre_signatures_in_state`. If the
flag is disabled, the functionality of pairing signature requests with
pre-signatures continues as before:
1. Consensus delivers the "view" of available pre-signatures that are
currently stored on the blockchain to execution in every round.
2. Execution removes all pre-signatures that were already matched from
this delivered "view".
3. Execution pairs the remaining pre-signatures with unmatched signature
requests, up to a maximum.

If the flag is enabled, delivered pre-signatures are instead stored in
the pre-signature stash, and paired with incoming requests from there:
1. Consensus delivers newly available pre-signatures only once (they are
no longer stored on the blockchain). Additionally, the key transcript is
delivered.
2. If the delivered key transcript is different to the one stored in the
stash, the stash is purged (old pre-signatures should not be used
anymore after a key rotation).
3. Delivered pre-signatures are merged into the pre-signature stash of
the corresponding key.
4. Signature requests are paired by popping pre-signatures from the
pre-signature stash.

This PR implements steps 2.-4. A feature flag is required because
> 1. Consensus delivers newly available pre-signatures only once (they
are no longer stored on the blockchain).

isn't implemented yet. Once implemented, this functionality should be
activated with the same release.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>