feat: Move disk encryption logic to Rust and support SEV-based disk encryption (#6170)
NODE-1663

This PR replaces the existing `cryptsetup luksOpen` and `cryptsetup
luksFormat` calls with a Rust binary that supports SEV-based disk
encryption.

- Add a new os tool called `guest_disk` which support `crypt-open` and
`crypt-format` subcommands
- The tool implements the current static key generation method when SEV
is off (implemented in `generated_key.rs`).
- Key generation has been moved into the tool as well and the
corresponding `generate-store-key.service` has been removed. This way
the key generation is covered by unit tests which was not the case
previously.
- The tool also supports disk encryption keys derived from SEV
measurements when SEV is enabled (implement in `sev.rs`).
- The tool already tries to unlock partitions using the key from the
`var` partition when a key is available
(`setup_store_with_previous_key`). Currently that file is never written,
this will be added in a later PR when we add the GuestOS upgrade logic.
- There is extensive test coverage for both the generated key and the
SEV codepath.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>
Co-authored-by: IDX GitHub Automation <>