feat: [EXC-2090] Add snapshot source to canister history (#6163)
Since the addition of snapshot uploading, a field `SnapshotSource` is
associated with each snapshot, distinguishing between snapshots taken
from a canister (impossible to tamper with) and a manually uploaded
snapshot (potentially manipulated by the uploader). These cases should
be distinguished in the canister history so that audits can take this
information into account.