feat(ICP-Rosetta): FI-1734: adding disburse maturity endpoint (#5994)
The functionality is similar to existing `disburse` endpoint. Instead of
disbursing a neuron, we disburse the maturity associated with the
neuron. We also provide the `recipient` account identifier. Instead of
providing the `amount` we provide the `percentage_to_disburse`, which
specifies the percentage of the maturity that will be disbursed.

---------

Co-authored-by: Maciej Dfinity <maciej@dfinity.org>