refactor: CON-1558 Remove `ThresholdSigInputsRef` (#6157)
Creating a tECDSA/tSchnorr signature requires certain inputs, for
instance the message (hash) to be signed, a nonce, and a pre-signature.
These inputs are derived from the request context, using the function
`build_signature_inputs`.

Previously, the request context only contained the pre-signature ID,
while the actual pre-signature remained on the blockchain.
Pre-signatures are composed of IDkgTranscripts. Since these transcripts
are comparatively large, they are only stored in a single block. The
latest block instead contains `PreSignatureRef`s which point to the
heights of the blockchain at which the real transcripts are stored.

Therefore, in order to build the complete inputs, two layers of
indirection had to be traversed:
1. Find the `PreSignatureRef` in latest finalized block, that
corresponds to the pre-signature ID of the request context
2. Using the `PreSignatureRef`, assemble a `ThresholdSigInputsRef`
struct which also contains the ref to the key transcript
3. "Resolve" the `ThresholdSigInputsRef` by looking up the real
transcripts in their respective blocks on the blockchain and assemble
the final `ThresholdSigInputs` struct.

However, since https://github.com/dfinity/ic/pull/6049, the request
context now contains the full pre-signature and key transcripts
directly. Consequently, this PR removes the indirection mentioned above,
and assembles the `ThresholdSigInputs` directly using only the data
stored in the context.

This greatly simplifies testing, as previously we had to prepare both
the contexts in state and the latest block payload to contain the
correct pre-signature IDs and pre-signature refs respectively.
Additionally, some failure test cases can be eliminated entirely, as
they can no longer happen by construction. For instance:
1. Failure to resolve a transcript during signature share
creation/validation/aggregation
2. Failure to find a pre-signature ref in the block for a given
pre-signature ID in the context
3. A context was paired with a pre-signature of the wrong scheme

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>