chore: CON-1560 Implement `ErrorReproducibility` for `BuildSignatureInputsError` (#6256)
Technically, all current `BuildSignatureInputsError`s _are_
reproducible, so the existing code was correct. But in principal, a new
non-reproducible variant may be added at any time. Therefore,
implementing `ErrorReproducibility` properly is a good idea.

We additionally add some tests for the case that the state ever contains
a malformed request context. In that case:
- Blocks containing signature agreements for a malformed request context
should be invalidated
- The IDKG client should not create signature shares for a malformed
request context
- Signature shares received for malformed request context are deferred

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>