test(node): Use dev hashes in mainnet-icos-revisions.json (#6261)
NODE-1684

The mainnet-icos-revisions.json hash values are all for prod images, but
we now support (and use!) dev images in our tests.

We got lucky (or unlucky) that the change to start using dev mainnet
images in tests didn’t break any of our tests. This was because we don’t
have any tests upgrading **_TO a mainnet version_**. Our tests only test
upgrading **FROM** mainnet versions or test ICs initially deployed with
mainnet versions.

If a test tried to upgrade to a mainnet version, it would fail because
the node would download a _dev_ upgrade image and see it doesn’t match
the expected _prod_ image hash.

I only noticed this when I was testing a node joining a testnet IC with
an unassigned registry version that didn’t match the node's own.