feat(ckbtc): reimburse withdrawals (#6272)
Add the ability to reimburse withdrawal requests which can be safely
canceled (future PR), meaning that

1. Either the minter never issued a Bitcoin transaction including that
withdrawal request;
2. Or it's guaranteed that such a transaction is no longer valid because
some of its UTXOs have been used by another transaction that is
considered finalized in the meantime.

`CI_OVERRIDE_DID_CHECK`: The PR introduced new events that break the
Candid API of the `get_events` endpoint. This is ok because this is a
debug endpoint and no other endpoint should be impacted.

---------

Co-authored-by: Thomas Locher <thomas.locher@dfinity.org>