feat: bootstrap NNS governance and root canisters (#6254)
This PR adds a new ICP feature `nns_governance` bootstrapping the NNS
governance ("@nns_governance_canister_test_at_mainnet_commit//file") and
the NNS root ("@mainnet_nns_root-canister//file") canisters.

This PR also fixes the implementation of the `cycles_minting` feature.
In particular, if the `cycles_minting` feature is requested, then the
default timestamp of a PocketIC instance is set to 10 May 2021 10:00:01
AM CEST (the smallest value that is strictly larger than the timestamp
in the CMC state).

Here's the current roadmap of ICP features in PocketIC (to cover the
functionality of the dfx nns extension):
[x] registry (NNS registry canister)
[x] cycles_minting (NNS cycles minting canister)
[x] icp_token (ICP ledger and index canisters)
[x] cycles_token (cycles ledger and index canisters)
[x] nns_governance (NNS governance and root canisters)  <-- this PR
[] sns (SNS-W and aggregator canisters)
[] nns_ui (NNS frontend dapp)
[] identity (Internet Identity canister)