fix: colocate api_bn_update_workload_test to reduce flakiness (#6148)
The `//rs/tests/boundary_nodes:api_bn_update_workload_test` has a high
flakiness rate of almost 5%. This test is the only test that uses the
[`ic_networking_subnet_update_workload`](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/tests/networking/subnet_update_workload/src/lib.rs?L121)
framework which is not colocated. For any test generating load from the
test-driver to the testnet it's important the test-driver is colocated
with the testnet to reduce the chance of cross DC traffic.