feat(ckbtc): cancel invalid stuck transactions (#6271)
If a transaction created during the withdrawal process (retrieve_btc)
has too many inputs, it is treated as invalid and user requests are
immediate refunded.

If such an invalid transaction is already in stuck state, the minter
will create and send a cancellation transaction to spend one of its
input utxos (and send to the minter itself). When this cancellation
transaction is finalized, user's withdrawal requests corresponding to
the previous stuck transaction will be refunded (minus a fee).

A small caveat is that existing retrieve_btc_status methods will not be
able to return a status reflecting either this error or refund status
without breaking backward compatibility. This will be addressed in a
future PR.

`CI_OVERRIDE_DID_CHECK`: The PR introduced new events that break the
Candid API of the get_events endpoint. This is ok because this is a
debug endpoint and no other endpoint should be impacted.

---------

Co-authored-by: Thomas Locher <thomas.locher@dfinity.org>
Co-authored-by: gregorydemay <112856886+gregorydemay@users.noreply.github.com>
Co-authored-by: gregorydemay <gregory.demay@dfinity.org>