feat(btc): initial support for dogecoin in ic-btc-adapter (#5822)
This is a first step towards a supporting dogecoin in ic-btc-adapter.
Notable changes are:

1. Point rust bitcoin version to
https://github.com/dfinity/rust-dogecoin, which supports both bitcoin
and dogecoin. At the moment this is branched off the same 0.35.5 version
used by ic-btc-canister, so there shouldn't be any breakage.
2. Because in this version `NetworkMessage` is parameterized by block
type, the implementation of ic-btc-adapter has to be adapted to this
change.
3. A new `AdapterNetwork` type is introduced in ic-btc-adapter to select
between bitcoin and dogecoin network types.
4. Address the difference between bitcoin and dogecoin in some
parameters used by ic-btc-adapter.
5. Existing tests for ic-btc-adapter still only test bitcoin, and
dogecoin related tests will be added in a separate PR.

The purpose of this PR is to switch to using rust-dogecoin crate, and to
add support for dogecoin in ic-btc-adapter's configuration without
changing its existing functionalities in supporting bitcoin. As a
result, the adapter now can download dogecoin blocks if run with a
dogecoin config, but full dogecoin support and tests will only be
provided in followup PRs.

---------

Co-authored-by: Thomas Locher <thomas.locher@dfinity.org>
Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>