chore(EXC-2098): Reorganize instrumentation (#6273)
This change is only moving code around so that the diff when
transitioning to `wirm` from `wasm_transform` is readable. In
`instrumentation.rs` we move the order of some functions around without
changing any of the bodies. In `system_api_replacements.rs` we modify
the indentation level to match the indentation that we will have after
the transition.