chore(idkg): CON-1564 Remove `IDkgBlockReader` from `IDkgBouncer`  (#6307)
When computing the new bouncer args, we are only interested in the
latest finalized height. Currently, this height is determined by cloning
the entire finalized chain cache, passing it to a new `IDkgBlockReader`
and calling the `tip_height()` function.

In this PR we simplify the process, by instead passing the
`ConsensusPoolCache` to the bouncer factory, which allows determining
the latest finalized height without cloning the entire chain.