feat(PocketIC): enable specifying initial time during PocketIC instance creation (#6315)
This PR enables specifying the initial timestamp of a newly created
PocketIC instance or if the new instance should make progress
automatically (i.e., if PocketIC should periodically update the time of
the instance to the real time and execute rounds on the subnets) during
PocketIC instance creation (instead of only after the instance is
created using a separate PocketIC operation).

The motivation for this PR is two-fold:
- setting time only after PocketIC instance creation is longer
equivalent to setting time during PocketIC instance creation if system
canisters are deployed during PocketIC instance creation (using the
`icp_features` option) - the system canisters could then see time that
is significantly off the time set afterwards that could be unexpected
for the system canisters (and they indeed produce canister logs related
to that);
- enabling auto progress during PocketIC instance creation ensures that
canister http outcalls are handled automatically by PocketIC and thus
system canisters (e.g., II to be added in a follow-up PR) could rely on
canister http outcalls.