fix: ensure head_nns variants of long system-tests only run once on nightly & introduce enable_head_nns_variant (#6321)
Currently callers of `system_test_nns` have to manually configure the
`extra_head_nns_tags` which determines the tags of the `head_nns`
variant of the system-test. This is error-prone as witnessed by the
following two bugs:

1. `//rs/tests/boundary_nodes:api_bn_integration_test` would override
the default `extra_head_nns_tags` with `[]` which caused the `head_nns`
variant to run on PRs which is not what we want. This is fixed by
https://github.com/dfinity/ic/pull/6318.
2.
[`//rs/tests/boundary_nodes:api_bn_update_workload_test`](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/tests/boundary_nodes/BUILD.bazel?L25)
which is marked as a `long_test` but `extra_head_nns_tags` is not
overridden with `[]` causing the `head_nns` variant to run both on
"Release Testing / CI Main / Bazel Test All" and "Release Testing /
Release System Tests" which is redundant and wastes CI resources. Not
fixed yet but will be fixed by this commit.

So this commit removes the `extra_head_nns_tags` argument of
`system_test_nns` and sets it automatically so users can't make the
above mistakes anymore.

A new `enable_head_nns_variant = True` argument has been introduced to
allow callers to disable the `head_nns` variant which is needed for a
number of tests.

An alternative design was considered in
https://github.com/dfinity/ic/pull/6320 but I like the simplicity of not
allowing setting `extra_head_nns_tags`.

Future Work
===

There's an idea to replace the `system_test_nns` macro with adding
`head_nns` targets explicitly using bazel transition rules. But that's a
bigger change which we have to consider carefully so until we have that
this simplification would prevent bugs like the ones mentioned above.