chore: do not allow subnet ID as callee in ingress messages (#6181)
This PR forbids to use a subnet ID as the callee in an ingress message
(to the management canister): the management canister ID must be used as
the callee instead. This leads to code simplification without breaking
any existing flows (see below for a justification).

On the ICP mainnet, ingress messages with a subnet ID as the callee are
always rejected anyway (and thus this PR is not a breaking change). This
is because such a call would also have to have that subnet ID as the
effective canister ID (enforced by the HTTP handler - see a new system
test in this PR) and then they would be rejected by a BN (they are
rejected in the execution environment in the new system test in this PR
which sends the ingress message to a replica directly bypassing a BN).