feat(XC): Support for Dogecoin in replica (#6257)
This PR adds support for Dogecoin to the Adapter client and context.

The PR adds three new variants to the `Network` tag of the internal
bitcoin interface.
It additionally adds the internal mechanisms that allow to specify
dogecoin adapters, to which the new calls can be routed.

This change should be sufficient to support dogecoin from the replica
site.