feat: [NODE-1686] Speed up perf tests (#6335)
This saves about 43 minutes when running the io perf tests. Most of the
time was spent writing zeros moving the GuestOS image onto the SSDs.

---

When GuestOS first boots, it will create a new encrypted volume with all
of the remaining space on disk. We already provide the entire
`hostlvm/guest` to the VM, so instead of writing/seeking through the
entire 5TiB of zeroes, we only need to clear signatures and write the
image to the front of the drive.