test: bump timeout of the //rs/tests/consensus/upgrade:upgrade_downgrade_nns_subnet_test (#6356)
The `//rs/tests/consensus/upgrade:upgrade_downgrade_nns_subnet_test`
frequently runs into the 20min timeout of the test task. It seems it's
not stuck at the point of timeout but just taking a long time so before
investigating further let's bump the timeout with an extra 10 minutes to
see if that fixes the flakiness.