feat: [CON-1487] introduce `NNSDelegationReader` (#6172)
# Preamble
The PR is big (sorry about that!) but the vast majority of the added
code are tests / benchmarks.

# Summary

See also https://github.com/dfinity/ic/pull/5907

Currently, the `delegations` we attach to responses to `/api/v3/call`,
`/api/v2/query`, `/api/v2/canister/read_state`, and
`/api/v2/subnet/read_state` requests, always have the same format:
```
tree
│
│
└─── time
│  
│  
└─── subnet
       │
       │
       └─── <subnet_id> 
                │
                │
                └─── public_key [cbor encoded key]
                │
                │
                └─── canister_ranges [cbor encoded list of canister ranges]
```
With the canister migration feature, the routing table might become very
fragmented, and while today each subnet has at most two canister ranges
assigned to it, in the feature that number might grow to thousands,
which means that in the current format the delegation might become very
large which would make it infeasible to attach it to every http
response.

To counteract this, we are going to introduce new endpoints, where we
are going to attach `delegations` in a format which contains a bounded
(currently by `5`) number of canister ranges which would suffice to
prove that a canister id is assigned to the subnet.
 
For `/api/v4/call`, `/api/v3/query`, and `/api/v3/canister/read_state`
endpoints the `delegations` will have the following format:
```
tree
│
│
└─── time
│ 
│  
└─── subnet
│      │
│      │
│      └─── <subnet_id> 
│               │
│               │
│               └─── public_key [cbor encoded key]
│
│
│
└─── canister_ranges
       │
       │
       └─── <subnet_id> 
                │
                │
                └─── <some_canister_id>
                            │
                            │
                            └─── canister_ranges [cbor encoded BOUNDED list of canister ranges]
```
and for `/api/v3/subnet/read_state`, we won't include any canister
ranges at all:
```
tree
│
│
└─── time
│ 
│  
└─── subnet
      │
      │
      └─── <subnet_id> 
               │
               │
               └─── public_key [cbor encoded key]
```

This PR introduces a `NNSDelegationBuilder` which builds nns delegations
in one of the three formats, but currently all the endpoints still
return the existing format.

# Changes
* introduced `NNSDelegationBuilder` - a struct which given a state tree,
builds an appropriate delegation
* introduced `NNSDelegationReader` as a wrapper around
`tokio::sync::watch::Reader` which makes it a bit easier to work with
* changed `verify_delegation_certificate` in `ic-certification` to
account for all formats of the delegations
* wrote some benchmarks

# Benchmark results
* On my devenv it took on average `73.245 µs` to retrieve a delegation
in the OLD format with `120_000` canister ranges assigned to the subnet.
The size of the delegation was `2'760'327` bytes.
* On my devenv it took on average `961.33 µs` to retrieve a delegation
in the NEW format with `120_000` canister ranges assigned to the subnet.
The size of the delegation was `1'031` bytes.

---------

Co-authored-by: Stefan Schneider <stefan.schneider@dfinity.org>