feat(ICP-Ledger): FI-1830: consent message for legacy transfer (#6357)
Dapps that need to send ICP to a recipient where they only know the
account id, e.g. a dex or so, need to use the `transfer` method. But
this method does not support icrc-21, so they can't do it from a
connected wallet.
This PR adds support for generic transfer to ICRC-21. The generic
transfer can then be used to provide ICRC-21 functionality to the legacy
ICP `transfer` endpoint.

---------

Co-authored-by: Mathias Björkqvist <mathias.bjorkqvist@dfinity.org>