fix: Use cached block chain to create `IDkgBlockReader` (#6287)
Whenever we need to create/validate/deliver an IDkg payload, we
instantiate the `IDkgBlockReader`. This is just a wrapper around a
`ConsensusBlockchain` (which is a wrapper around a `BTreeMap`)
containing all blocks on the chain up until the parent of the payload
being created/validated/delivered.

This reader is required for IDkg, in order to efficiently look up blocks
containing `IDkgTranscript`s that correspond to specific
`TranscriptRef`s.

Currently, the underlying `BTreeMap` is created from scratch by
deserializing blocks from the validated pool every time. With this PR,
we instead make use of the `CachedChainIterator` to instantiate this
map, which will first try to find the blocks in the pool's finalized
chain cache.

Only if the block cannot be found in the finalized chain cache (i.e.
because it hasn't been finalized yet), then it is looked up in the
validated pool.

Additionally, we fix a minor bug in the `CachedChainIterator`: When
trying to find the parent of a given block, the iterator first checks
the finalized chain, and then the validated proposals in the pool.
However, we should additionally consider the case where the parent is
the summary block, which may only exist as part of the catch up package,
in case the replica just started.

This wasn't a problem until now, because the `CachedChainIterator`
additionally takes an optional `to_block` as an argument, and all
callers set it to `Some(cache.catch_up_package().content.block)`, which
is exactly the block that would be missing if the argument were set to
`None`.

Finalization rate before (green is tECDSA signing subnet):
<img width="669" height="391" alt="image"
src="https://github.com/user-attachments/assets/3f91ff92-b9da-484a-954a-fe3a6934c0c6"
/>
The finalization rate drops towards the end of the interval, likely due
to more blocks being deserialized during every new block
creation/validation

Finalization rate after (yellow is tECDSA signing subnet):
<img width="506" height="386" alt="image"
src="https://github.com/user-attachments/assets/1495553b-4f15-4350-83ba-fa6cb791213d"
/>
The finalization rate is more stable (apart from the CUP)