feat: bootstrap Internet Identity canister (#6288)
This PR adds a new ICP feature `ii` bootstrapping the Internet Identity
(dev) canister ("@internet_identity_dev_at_mainnet_commit//file").

This PR disables the archiving feature of the Internet Identity.

This PR also adds CI checks that the hard-coded type of the Internet
Identity initial argument matches the actual type specified in the
corresponding custom section of the canister WASM.

Here's the current roadmap of ICP features in PocketIC (to cover the
functionality of the dfx nns extension):
[x] registry (NNS registry canister)
[x] cycles_minting (NNS cycles minting canister)
[x] icp_token (ICP ledger and index canisters)
[x] cycles_token (cycles ledger and index canisters)
[x] nns_governance (NNS governance and root canisters)
[x] sns (SNS-W and aggregator canisters)
[x] identity (Internet Identity canister) <-- this PR
[] nns_ui (NNS frontend dapp)