chore: rename main_memory to wasm_memory in snapshot data API (#6420)
This PR renames `main_memory` to `wasm_memory` in the snapshot data API
of the management canister. Using the term `wasm_memory` to refer to the
WASM (a.k.a. heap) memory is consistent with `wasm_memory_size` in the
snapshot metadata API and with `wasm_memory_limit` and
`wasm_memory_threshold` in the `canister_status` endpoint of the
management canister.