perf: Replace `RoutingTable::route` by `NetworkTopology::route` (#6361)
This commit removes `RoutingTable::route`, which was used to route
messages to the correct subnet.

The difference to the still existing `RoutingTable::lookup_entry` was
that `route` would additionally consider the case where the target is a
subnet id instead of a canister id. However, previously this was
implemented by scanning over all canister ranges and checking if the
input occurs as the subnet id for any range. There are two issues with
this, which this commit both solves:

1. As it's a linear scan, route was `O(n)`, where n is the number of
ranges in the routing table. Currently `n` is small, but will grow with
the upcoming canister migration feature.

2. There is a theoretical edge case of subnets with no ranges assigned
to them. It was not possible to route to such a subnet.

This commit solves both problems by instead implementing `route` in the
`NetworkTopology` struct. `NetworkTopology` contains both a
`RoutingTable` and a `BTreeMap` of subnets. Instead of a linear scan, we
now only need to do a quick lookup in the `BTreeMap`. Furthermore,
subnets without any ranges will also be found this way.

This commit is somewhat large due to the necessary test changes. In
production code, the switch is very simple.