chore(Ledger_archives): No special COMPRESSED_CANISTERS treatment for ledger archives (#6440)
Move the ICP and ICRC ledger archive canisters from the
`COMPRESSED_CANISTERS` list into the regular `CANISTERS` for building
and publishing canisters. Since all canisters are compressed these days,
it doesn't seem necessary to have the ledger archives in the separate
list, and having them in the `CANISTERS` list also means that their
`.did` files are also bundled and published.