refactor(XC): More consistent naming of native cross chain code (#6383)
This PR renames some structures and enums around the Bitcoin feature.
Since they are no longer bitcoin specific with the integration of
dogecoin, it makes sense to rename them.

This PR does not attempt to rename everything in one go. In particular,
it does not touch file/crate names, consensus internals or the
`ic_management_canister_types_private` exports.