chore: refactor SubnetAvailableMemory (#6447)
This PR refactors `SubnetAvailableMemory` by splitting its constructor
`new` into a pair of constructors:
- `new_for_testing`: no scaling factor applied in tests that create
`SubnetAvailableMemory` manually (only to be used in tests);
- `new_scaled`: used internally in the execution environment to apply
the scaling factor `self.scheduler_cores` (renamed from
`self.resource_saturation_scaling` to make the semantics of the value
clearer);

removing the division operator on `SubnetAvailableMemory` (an instance
of `SubnetAvailableMemory` should be created with an appropriate scaling
factor and not rescaled later) and the scaling factor as an explicit
field of `SubnetAvailableMemory` (since `SubnetAvailableMemory` is
always scaled with the same scaling factor).

This way, we scale `SubnetAvailableMemory` with the same scaling factor
as `ResourceSaturation` by design (instead of relying on the fact that
the same value of the scaling factor is used at the moment when
computing `ResourceSaturation` and when manually scaling
`SubnetAvailableMemory`).

This PR also renames `subnet_memory_capacity` to
`full_subnet_memory_capacity` to make it explicit that no scaling is
applied.

Finally, existing tests are updated.