feat: CON-1571 Count the total number of available pre-signatures in the state and blockchain (#6446)
This PR adds a function to count the total number of available
pre-signatures that are stored in the state at a given height, and all
payloads above the state height.

The function is currently unused. In the future it will be needed to
determine the number of additional pre-signature creations that we
should start for each stash.

Counting the number of pre-signatures correctly additionally relies on
https://github.com/dfinity/ic/pull/6308, which ensures that all blocks
above the certified height are part of the `IDkgBlockReader`.