chore(PocketIC): turn nonmainnet_features into struct (#6442)
This PR turns the field `nonmainnet_features` in the argument of the
PocketIC server endpoint `/instances/` to create a new PocketIC instance
into an optional struct field specifying which features not available on
the ICP mainnet should be available in the newly created PocketIC
instance (the field `nonmainnet_features` used to be a simple Boolean
before).

The vetKD feature has been rolled out to the ICP mainnet in the meantime
and thus there's no need to enable beta features to have vetKD enabled.